requirements.txt
python-dotenv
google-generativeai
langchain
PyMuPDF
sentence-transformers
transformers
faiss-cpu
